backend-django/requirements.txt
Django>=4.1.7,<4.2
djangorestframework>=3.14.0,<3.15
django-cors-headers>=3.13.0,<3.14
django-environ>=0.9.0,<0.10
numpy>=1.24.2,<1.25
matplotlib>=3.7.0,<3.8
pymysql>=1.0.2,<1.1
pandas>=1.5.3,<1.6
polars>=0.16.1,<0.17
seaborn>=0.11.2,<0.12
statannotations @ git+https://github.com/trevismd/statannotations.git@dev

plotly>=5.13.1,<5.14
dash-bio>=1.0.2,<1.1

pyarrow>=11.0.0,<11.1

statsmodels>=0.14.0,<0.15
